Update all submodules to track master, add/remove a few
This commit uses git 1.8.2 feature to track master of submodules
instead of a specific commit, so that cloning this repo will
automatically pull in the latest version of things.

It also removes the neco-ghc plugin for vim since I'm not planning on
learning Haskell anymore and adds the arm assembly (armasm) and ansi
highlighting (AnsiEsc) plugins as git submodules.